Use stop signal channel to stop worker immediately
Use channel for stop signal instead of checking for c.workerIsRunning.
It is because that if the toggle is off, the msg := <-c.messageQueue is still
waiting for incoming message and will be able to send one last message even
after StopWorker() is invoked.
simply put, StopWorker() does not immediately kill the worker goroutine.